+++ modules/module_a/tools/__init__.py
@@ -1,3 +1,4 @@
 from .math import add, subtract
+from .fetch_external import fetch_external_data
 
-__all__ = ['add', 'subtract']
+__all__ = ["add", "subtract", "fetch_external_data"]